[Fix] Invaild type errorCode(Int64 -> Int)

+++ ios/Classes/model/error/AuthenticatedFailed.swift
@@ -1,8 +1,8 @@
 internal struct AuthenticatedFailed : BaseError {
-    var errorCode: Int64
+    var errorCode: Int
     var message: String?
 
-    init (errorCode: Int64, message: String) {
+    init (errorCode: Int, message: String) {
         self.errorCode = errorCode
         self.message = message
     }
